Move app detail pane tests to their own file.
