+++ __tests__/AppDetailPane.js
@@ -1,0 +1,422 @@
+import '@testing-library/jest-dom/extend-expect';
+
+import { fireEvent, waitFor } from '@testing-library/react';
+import RoutePath from 'lib/routePath';
+import { getInstallationInfo } from 'model/services/giantSwarm';
+import nock from 'nock';
+import { StatusCodes } from 'shared/constants';
+import { OrganizationsRoutes } from 'shared/constants/routes';
+import {
+  API_ENDPOINT,
+  appCatalogsResponse,
+  appResponseWithCustomConfig,
+  appsResponse,
+  AWSInfoResponse,
+  catalogIndexResponse,
+  getMockCall,
+  getMockCallTimes,
+  ORGANIZATION,
+  orgResponse,
+  orgsResponse,
+  releasesResponse,
+  userResponse,
+  V4_CLUSTER,
+  v4AWSClusterResponse,
+  v4AWSClusterStatusResponse,
+  v4ClustersResponse,
+} from 'testUtils/mockHttpCalls';
+import { renderRouteWithStore } from 'testUtils/renderUtils';
+
+describe('Installed app detail pane', () => {
+  beforeEach(() => {
+    getInstallationInfo.mockResolvedValueOnce(AWSInfoResponse);
+    getMockCall('/v4/clusters/', v4ClustersResponse);
+    getMockCallTimes('/v4/organizations/', orgsResponse);
+  });
+
+  it('updates the config map of an already installed app', async () => {
+    nock('https://catalogshost')
+      .get('/giantswarm-catalog/index.yaml')
+      .reply(StatusCodes.Ok, catalogIndexResponse);
+
+    nock(API_ENDPOINT)
+      .intercept(`/v4/clusters/${V4_CLUSTER.id}/apps/my%20app/config/`, 'PATCH')
+      .reply(StatusCodes.Ok);
+
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/apps/`,
+      [appResponseWithCustomConfig],
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText, getByText } = renderRouteWithStore(clusterDetailPath);
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    // Click on app to open the editing modal
+    const appLabel = getByText(/chart version: 0.0.1/i);
+    fireEvent.click(appLabel);
+
+    // Delete the existing file
+    const fileInputPlaceholder = getByText(/configmap has been set/i);
+    const fileInput = fileInputPlaceholder.parentNode.querySelector('input');
+    const file = new Blob(
+      [
+        JSON.stringify({
+          name: 'test',
+          namespace: 'some-other-test',
+        }),
+      ],
+      {
+        type: 'application/json',
+        name: 'config2.json',
+      }
+    );
+
+    fireEvent.change(fileInput, {
+      target: {
+        files: [file],
+      },
+    });
+
+    await findByText(/has successfully been updated./i);
+  });
+
+  it('deletes the config map of an already installed app', async () => {
+    nock('https://catalogshost')
+      .get('/giantswarm-catalog/index.yaml')
+      .reply(StatusCodes.Ok, catalogIndexResponse);
+
+    nock(API_ENDPOINT)
+      .intercept(
+        `/v4/clusters/${V4_CLUSTER.id}/apps/my%20app/config/`,
+        'DELETE'
+      )
+      .reply(StatusCodes.Ok);
+
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/apps/`,
+      [appResponseWithCustomConfig],
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText, getByText, queryByText } = renderRouteWithStore(
+      clusterDetailPath
+    );
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    // Click on app to open the editing modal
+    const appLabel = getByText(/chart version: 0.0.1/i);
+    fireEvent.click(appLabel);
+
+    // Upload a configmap file
+    const fileInputPlaceholder = getByText(/configmap has been set/i);
+    let deleteButton = fileInputPlaceholder.parentNode.querySelector(
+      '.btn-danger'
+    );
+    fireEvent.click(deleteButton);
+
+    // Confirm deletion
+    deleteButton = getByText(/^delete configmap$/i);
+    fireEvent.click(deleteButton);
+
+    await waitFor(() => {
+      expect(queryByText(/delete configmap/i)).not.toBeInTheDocument();
+    });
+
+    await findByText(/has been deleted./i);
+  });
+
+  it('updates secrets of an already installed app', async () => {
+    nock('https://catalogshost')
+      .get('/giantswarm-catalog/index.yaml')
+      .reply(StatusCodes.Ok, catalogIndexResponse);
+
+    nock(API_ENDPOINT)
+      .intercept(`/v4/clusters/${V4_CLUSTER.id}/apps/my%20app/secret/`, 'PATCH')
+      .reply(StatusCodes.Ok);
+
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/apps/`,
+      [appResponseWithCustomConfig],
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText, getByText, queryByText } = renderRouteWithStore(
+      clusterDetailPath
+    );
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    // Click on app to open the editing modal
+    const appLabel = getByText(/chart version: 0.0.1/i);
+    fireEvent.click(appLabel);
+
+    // Upload a secrets file
+    const fileInputPlaceholder = getByText(/secret has been set/i);
+    const fileInput = fileInputPlaceholder.parentNode.querySelector('input');
+    const file = new Blob(
+      [
+        JSON.stringify({
+          name: 'test',
+          namespace: 'some-other-test',
+        }),
+      ],
+      {
+        type: 'application/json',
+        name: 'secret.json',
+      }
+    );
+    fireEvent.change(fileInput, {
+      target: {
+        files: [file],
+      },
+    });
+
+    await waitFor(() => {
+      expect(queryByText(/delete secret/i)).not.toBeInTheDocument();
+    });
+
+    await findByText(/has successfully been updated./i);
+  });
+
+  it('deletes secrets of an already installed app', async () => {
+    nock('https://catalogshost')
+      .get('/giantswarm-catalog/index.yaml')
+      .reply(StatusCodes.Ok, catalogIndexResponse);
+
+    nock(API_ENDPOINT)
+      .intercept(
+        `/v4/clusters/${V4_CLUSTER.id}/apps/my%20app/secret/`,
+        'DELETE'
+      )
+      .reply(StatusCodes.Ok);
+
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse);
+    getMockCall(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse
+    );
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/apps/`,
+      [appResponseWithCustomConfig],
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText, getByText, queryByText } = renderRouteWithStore(
+      clusterDetailPath
+    );
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    // Click on app to open the editing modal
+    const appLabel = getByText(/chart version: 0.0.1/i);
+    fireEvent.click(appLabel);
+
+    // Delete the existing file
+    const fileInputPlaceholder = getByText(/secret has been set/i);
+    let deleteButton = fileInputPlaceholder.parentNode.querySelector(
+      '.btn-danger'
+    );
+    fireEvent.click(deleteButton);
+
+    // Confirm deletion
+    deleteButton = getByText(/^delete secret$/i);
+    fireEvent.click(deleteButton);
+
+    await waitFor(() => {
+      expect(queryByText(/delete secret/i)).not.toBeInTheDocument();
+    });
+
+    await findByText(/has been deleted./i);
+  });
+
+  it('deletes already installed app', async () => {
+    nock('https://catalogshost')
+      .get('/giantswarm-catalog/index.yaml')
+      .reply(StatusCodes.Ok, catalogIndexResponse);
+
+    nock(API_ENDPOINT)
+      .intercept(`/v4/clusters/${V4_CLUSTER.id}/apps/my%20app/`, 'DELETE')
+      .reply(StatusCodes.Ok);
+
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCallTimes(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse, 2);
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse,
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    // Before deleting the app, there are apps.
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/apps/`, appsResponse);
+
+    // After deleting the app, there are no apps.
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/apps/`, []);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText, getByText, queryByText } = renderRouteWithStore(
+      clusterDetailPath
+    );
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    // Click on app to open the editing modal
+    const appLabel = getByText(/chart version: 0.0.1/i);
+    fireEvent.click(appLabel);
+
+    // Delete the app
+    let deleteButton = getByText(/delete app/i);
+    fireEvent.click(deleteButton);
+
+    // Confirm deletion
+    deleteButton = getByText(/delete app/i);
+    fireEvent.click(deleteButton);
+
+    await waitFor(() => {
+      expect(queryByText(/delete app/i)).not.toBeInTheDocument();
+    });
+
+    await findByText(/will be deleted/i);
+  });
+
+  it('shows a no apps installed message when there are no apps yet', async () => {
+    getMockCallTimes('/v4/user/', userResponse);
+    getMockCall(`/v4/organizations/${ORGANIZATION}/`, orgResponse);
+    getMockCall('/v4/appcatalogs/', appCatalogsResponse);
+    getMockCallTimes(`/v4/organizations/${ORGANIZATION}/credentials/`, [], 2);
+    getMockCallTimes(`/v4/clusters/${V4_CLUSTER.id}/`, v4AWSClusterResponse, 2);
+    getMockCallTimes(
+      `/v4/clusters/${V4_CLUSTER.id}/status/`,
+      v4AWSClusterStatusResponse,
+      2
+    );
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/apps/`, []);
+    getMockCall(`/v4/clusters/${V4_CLUSTER.id}/key-pairs/`);
+    getMockCall('/v4/releases/', releasesResponse);
+
+    const clusterDetailPath = RoutePath.createUsablePath(
+      OrganizationsRoutes.Clusters.Detail,
+      {
+        orgId: ORGANIZATION,
+        clusterId: V4_CLUSTER.id,
+      }
+    );
+    const { findByText } = renderRouteWithStore(clusterDetailPath);
+
+    const appsTab = await findByText(/^apps$/i);
+    fireEvent.click(appsTab);
+
+    expect(
+      await findByText('No apps installed on this cluster')
+    ).toBeInTheDocument();
+  });
+});
